+++ src/TestableHttpClient/Responses.cs
@@ -52,12 +52,6 @@ public static class Responses
     /// <returns>An HttpResponse with the configured StatusCode.</returns>
     public static IResponse StatusCode(HttpStatusCode statusCode) => new HttpResponse(statusCode);
     /// <summary>
-    /// Create a response with the NoContent status code.
-    /// </summary>
-    /// <returns>An HttpResponse with the configured StatusCode.</returns>
-    [Obsolete("Please use StatusCode(HttpStatusCode.NoContent) instead.")]
-    public static IResponse NoContent() => StatusCode(HttpStatusCode.NoContent);
-    /// <summary>
     /// Create a response with some text content.
     /// </summary>
     /// <param name="content">The content to put in the response.</param>